Module Geneweb_rpc.Service

type error = [

| Encoding.error

| `Bad_arity

]

val pp_error : error Fmt.t

type 'a res = ('a, error) Lwt_result.t

module Desc : sig ... end

type ('a, 'r) meth = private {

name : string;

desc : ('a, 'r) Desc.t;

f : 'a;

}

Type of method.

val decl : string -> ('a, 'r) Desc.t -> 'a -> ('a, 'r) meth

decl n desc f declares the method n with description desc and the underlying function f.

type t

Type of a service. A service is just a heterogeneous set of methods.

type binding = private

| Binding : ('a, 'r) Desc.t * 'a -> binding
(*
Internal binding used to store the methods.
*)

val empty : t

An empty service.

val add : ('a, 'r) meth -> t -> t

add meth s adds the method meth in the service s.

val find : string -> t -> binding option

find name s finds the method name in the service s.

val fold : (string -> binding -> 'a -> 'a) -> t -> 'a -> 'a

module PingPong : sig ... end

module Search : sig ... end