Module Service.Desc

type ('a, 'r) t

Type of the encoding of an arrow of type 'a. The parametric type r is the return type of the arrow.

val eval : ('a, _) t -> 'a -> Yojson.Safe.t list -> Yojson.Safe.t res

val arity : ('a, _) t -> int

val pp_desc : ('a, 'r) t Fmt.t

pp_desc ppf e printst the encoding e on the formatter ppf for debugging purposes.

module Syntax : sig ... end